mopra reader. changed hours to double (was: long). removed blockbasedfitnessfunction. bump to 0.5

+++ src/main/java/local/radioschedulers/Job.java
@@ -24,7 +24,7 @@ public class Job implements Comparable<Job> {
 	/**
 	 * total hours needed
 	 */
-	public Long hours;
+	public Double hours;
 
 	/**
 	 * start of LST time that can be used
