+++ src/main/java/local/radioschedulers/importer/SeedsGeneratingProposalReader.java
@@ -10,7 +10,7 @@ public class SeedsGeneratingProposalReader implements IProposalReader {
 
 	List<Proposal> proposals = new ArrayList<Proposal>();
 	
-	public Proposal createSimpleProposal(String name, double prio, double startlst, double endlst, Long totalhours) {
+	public Proposal createSimpleProposal(String name, double prio, double startlst, double endlst, double totalhours) {
 		Proposal p = new Proposal();
 		p.id = name;
 		p.name = name;
